[#18] Test required field validation for prefilled form fields.
In order to optimize the tests execution time, we only render the form once, in beforeAll. By default, react-testing-library calls a cleanup function (which unmounts React trees that were mounted with render) after each test. To prevent it, we import render function from testing-library/react/pure that doesn't call the cleanup function after each test, which means that we need to call the cleanup function manually.

[@aiwatko]