[#19] Persist logged in user to localStorage
A UX concern arose: When you are on the habitat use form and refresh the page, you are redirected to /login and back again. To combat this suboptimal behavior, we store the logged in user to localStorage and hydrate react state from it when the application loads. The firebase onAuthStateChanged event handler obviously changes localStorage and react state too, so that nothing gets out of sync.

[@peersmg / @rahul-rakshit]