[#43] Create component reload app when app updates.
Given that create-react-app doesn't let you modify workbox config, our
options are limited.

In our solution, we're activating the new service worker as soon as it becomes available. We also provide a button to allow the user to reload the app, which then allows the new service worker to respond with the new pre-cached content.

[@rahul-rakshit / @flaviarodriguesf]