[#72] Introduce TextInput with unit tests.
In this commit, we've created a test helper called renderWithinFormik,
which lets you test individual form fields at a unit-level. We've
created a TextInput field which we'll use for the area field of the
Encounter Form. In a later refactor, we can use this component to
replace all text input fields.

The encounter form now only contains a single TextInput field and it
will grow as we work on this story.

We've needed to update react testing library to the
newest version. This deprecates "wait", so other test suites have been
updated to the newer "waitFor" function. A shim called
"mutationobserver-shim" is used in the "setupTests" file to use
"waitFor" with "jest-dom".

[@rahul-rakshit / @peersmg]